LPS-96062 Apply changes
