+++ modules/apps/change-tracking/change-tracking-change-lists-configuration-web/src/main/java/com/liferay/change/tracking/change/lists/configuration/web/internal/portlet/ChangeListsConfigurationPortlet.java
@@ -14,7 +14,7 @@
 
 package com.liferay.change.tracking.change.lists.configuration.web.internal.portlet;
 
-import com.liferay.change.tracking.configuration.CTPortalConfiguration;
+import com.liferay.change.tracking.configuration.CTConfiguration;
 import com.liferay.change.tracking.constants.CTPortletKeys;
 import com.liferay.portal.configuration.metatype.bnd.util.ConfigurableUtil;
 import com.liferay.portal.kernel.model.Role;
@@ -46,7 +46,7 @@ import org.osgi.service.component.annotations.Modified;
  * @author Máté Thurzó
  */
 @Component(
-	configurationPid = "com.liferay.change.tracking.configuration.CTPortalConfiguration",
+	configurationPid = "com.liferay.change.tracking.configuration.CTConfiguration",
 	configurationPolicy = ConfigurationPolicy.OPTIONAL, immediate = true,
 	property = {
 		"com.liferay.portlet.css-class-wrapper=portlet-change-lists-configuration",
@@ -90,8 +90,8 @@ public class ChangeListsConfigurationPortlet extends MVCPortlet {
 	@Activate
 	@Modified
 	protected void activate(Map<String, Object> properties) {
-		_ctPortalConfiguration = ConfigurableUtil.createConfigurable(
-			CTPortalConfiguration.class, properties);
+		_ctConfiguration = ConfigurableUtil.createConfigurable(
+			CTConfiguration.class, properties);
 	}
 
 	@Override
@@ -106,7 +106,7 @@ public class ChangeListsConfigurationPortlet extends MVCPortlet {
 		}
 
 		String[] administratorRoleNames =
-			_ctPortalConfiguration.administratorRoleNames();
+			_ctConfiguration.administratorRoleNames();
 
 		UserBag userBag = permissionChecker.getUserBag();
 
@@ -122,6 +122,6 @@ public class ChangeListsConfigurationPortlet extends MVCPortlet {
 				permissionChecker.getUserId()));
 	}
 
-	private CTPortalConfiguration _ctPortalConfiguration;
+	private CTConfiguration _ctConfiguration;
 
 }
