+++ modules/apps/change-tracking/change-tracking-change-lists-history-web/src/main/java/com/liferay/change/tracking/change/lists/history/web/internal/portlet/ChangeListsHistoryPortlet.java
@@ -14,7 +14,7 @@
 
 package com.liferay.change.tracking.change.lists.history.web.internal.portlet;
 
-import com.liferay.change.tracking.configuration.CTPortalConfiguration;
+import com.liferay.change.tracking.configuration.CTConfiguration;
 import com.liferay.change.tracking.constants.CTPortletKeys;
 import com.liferay.change.tracking.constants.CTWebKeys;
 import com.liferay.change.tracking.model.CTCollection;
@@ -51,7 +51,7 @@ import org.osgi.service.component.annotations.Reference;
  * @author Laszlo Pap
  */
 @Component(
-	configurationPid = "com.liferay.change.tracking.configuration.CTPortalConfiguration",
+	configurationPid = "com.liferay.change.tracking.configuration.CTConfiguration",
 	configurationPolicy = ConfigurationPolicy.OPTIONAL, immediate = true,
 	property = {
 		"com.liferay.portlet.add-default-resource=false",
@@ -112,8 +112,8 @@ public class ChangeListsHistoryPortlet extends MVCPortlet {
 	@Activate
 	@Modified
 	protected void activate(Map<String, Object> properties) {
-		_ctPortalConfiguration = ConfigurableUtil.createConfigurable(
-			CTPortalConfiguration.class, properties);
+		_ctConfiguration = ConfigurableUtil.createConfigurable(
+			CTConfiguration.class, properties);
 	}
 
 	@Override
@@ -128,7 +128,7 @@ public class ChangeListsHistoryPortlet extends MVCPortlet {
 		}
 
 		String[] administratorRoleNames =
-			_ctPortalConfiguration.administratorRoleNames();
+			_ctConfiguration.administratorRoleNames();
 
 		UserBag userBag = permissionChecker.getUserBag();
 
@@ -147,6 +147,6 @@ public class ChangeListsHistoryPortlet extends MVCPortlet {
 	@Reference
 	private CTCollectionLocalService _ctCollectionLocalService;
 
-	private CTPortalConfiguration _ctPortalConfiguration;
+	private CTConfiguration _ctConfiguration;
 
 }
