+++ modules/apps/change-tracking/change-tracking-change-lists-history-web/src/main/resources/META-INF/resources/init.jsp
@@ -24,8 +24,8 @@ taglib uri="http://liferay.com/tld/ui" prefix="liferay-ui" %>
 
 <%@ page import="com.liferay.change.tracking.change.lists.history.web.internal.display.context.ChangeListsHistoryDetailsDisplayContext" %><%@
 page import="com.liferay.change.tracking.change.lists.history.web.internal.display.context.ChangeListsHistoryDisplayContext" %><%@
-page import="com.liferay.change.tracking.configuration.CTConfigurationRegistryUtil" %><%@
 page import="com.liferay.change.tracking.constants.CTWebKeys" %><%@
+page import="com.liferay.change.tracking.definition.CTDefinitionRegistryUtil" %><%@
 page import="com.liferay.change.tracking.model.CTCollection" %><%@
 page import="com.liferay.change.tracking.model.CTEntry" %><%@
 page import="com.liferay.petra.string.StringPool" %><%@
