+++ modules/apps/change-tracking/change-tracking-test/src/testIntegration/java/com/liferay/change/tracking/service/test/CTEngineManagerTest.java
@@ -15,11 +15,11 @@
 package com.liferay.change.tracking.service.test;
 
 import com.liferay.arquillian.extension.junit.bridge.junit.Arquillian;
-import com.liferay.change.tracking.configuration.CTConfiguration;
-import com.liferay.change.tracking.configuration.CTConfigurationRegistrar;
-import com.liferay.change.tracking.configuration.builder.CTConfigurationBuilder;
 import com.liferay.change.tracking.constants.CTConstants;
 import com.liferay.change.tracking.constants.CTPortletKeys;
+import com.liferay.change.tracking.definition.CTDefinition;
+import com.liferay.change.tracking.definition.CTDefinitionRegistrar;
+import com.liferay.change.tracking.definition.builder.CTDefinitionBuilder;
 import com.liferay.change.tracking.engine.CTEngineManager;
 import com.liferay.change.tracking.engine.CTManager;
 import com.liferay.change.tracking.model.CTCollection;
@@ -107,7 +107,7 @@ public class CTEngineManagerTest {
 		_testResourceClassName = _classNameLocalService.addClassName(
 			TestResourceModelClass.class.getName());
 
-		_ctConfiguration = _ctConfigurationBuilder.setContentType(
+		_ctDefinition = _ctDefinitionBuilder.setContentType(
 			"Test Object"
 		).setContentTypeLanguageKey(
 			"test-object"
@@ -133,13 +133,13 @@ public class CTEngineManagerTest {
 			testVersion -> WorkflowConstants.STATUS_APPROVED
 		).build();
 
-		_ctConfigurationRegistrar.register(_ctConfiguration);
+		_ctDefinitionRegistrar.register(_ctDefinition);
 	}
 
 	@After
 	public void tearDown() throws Exception {
-		if (_ctConfiguration != null) {
-			_ctConfigurationRegistrar.unregister(_ctConfiguration);
+		if (_ctDefinition != null) {
+			_ctDefinitionRegistrar.unregister(_ctDefinition);
 		}
 
 		// If the change tracking was enabled originally, then leave it in the
@@ -709,7 +709,7 @@ public class CTEngineManagerTest {
 
 	@Test
 	public void testIsChangeTrackingSupported() throws Exception {
-		_ctConfigurationRegistrar.unregister(_ctConfiguration);
+		_ctDefinitionRegistrar.unregister(_ctDefinition);
 
 		boolean changeTrackingSupported =
 			_ctEngineManager.isChangeTrackingSupported(
@@ -723,7 +723,7 @@ public class CTEngineManagerTest {
 
 		Assert.assertFalse(changeTrackingSupported);
 
-		_ctConfigurationRegistrar.register(_ctConfiguration);
+		_ctDefinitionRegistrar.register(_ctDefinition);
 
 		changeTrackingSupported = _ctEngineManager.isChangeTrackingSupported(
 			TestPropsValues.getCompanyId(), TestVersionModelClass.class);
@@ -939,14 +939,14 @@ public class CTEngineManagerTest {
 	@Inject
 	private CTCollectionLocalService _ctCollectionLocalService;
 
-	private CTConfiguration _ctConfiguration;
+	private CTDefinition _ctDefinition;
 
 	@Inject
-	private CTConfigurationBuilder
-		<TestResourceModelClass, TestVersionModelClass> _ctConfigurationBuilder;
+	private CTDefinitionBuilder<TestResourceModelClass, TestVersionModelClass>
+		_ctDefinitionBuilder;
 
 	@Inject
-	private CTConfigurationRegistrar _ctConfigurationRegistrar;
+	private CTDefinitionRegistrar _ctDefinitionRegistrar;
 
 	@Inject
 	private CTEngineManager _ctEngineManager;
