+++ modules/apps/dynamic-data-mapping/dynamic-data-mapping-change-tracking/src/main/java/com/liferay/dynamic/data/mapping/change/tracking/internal/DDMStructureCTRegistrar.java
@@ -12,23 +12,21 @@
  * details.
  */
 
-package com.liferay.dynamic.data.mapping.change.tracking.internal.configuration;
+package com.liferay.dynamic.data.mapping.change.tracking.internal;
 
-import com.liferay.change.tracking.configuration.CTConfigurationRegistrar;
-import com.liferay.change.tracking.configuration.builder.CTConfigurationBuilder;
+import com.liferay.change.tracking.definition.CTDefinitionRegistrar;
+import com.liferay.change.tracking.definition.builder.CTDefinitionBuilder;
 import com.liferay.change.tracking.function.CTFunctions;
-import com.liferay.dynamic.data.mapping.model.DDMTemplate;
-import com.liferay.dynamic.data.mapping.model.DDMTemplateVersion;
-import com.liferay.dynamic.data.mapping.model.DDMTemplateVersionModel;
-import com.liferay.dynamic.data.mapping.service.DDMTemplateLocalService;
-import com.liferay.dynamic.data.mapping.service.DDMTemplateVersionLocalService;
+import com.liferay.dynamic.data.mapping.model.DDMStructure;
+import com.liferay.dynamic.data.mapping.model.DDMStructureVersion;
+import com.liferay.dynamic.data.mapping.service.DDMStructureLocalService;
+import com.liferay.dynamic.data.mapping.service.DDMStructureVersionLocalService;
 import com.liferay.portal.kernel.dao.orm.DynamicQuery;
 import com.liferay.portal.kernel.dao.orm.Property;
 import com.liferay.portal.kernel.dao.orm.PropertyFactoryUtil;
 import com.liferay.portal.kernel.exception.PortalException;
 import com.liferay.portal.kernel.log.Log;
 import com.liferay.portal.kernel.log.LogFactoryUtil;
-import com.liferay.portal.kernel.service.GroupLocalService;
 import com.liferay.portal.kernel.util.LocaleUtil;
 import com.liferay.portal.kernel.workflow.WorkflowConstants;
 
@@ -45,65 +43,70 @@ import org.osgi.service.component.annotations.ReferenceScope;
  * @author Máté Thurzó
  */
 @Component(immediate = true, service = {})
-public class DDMTemplateCTConfigurationRegistrar {
+public class DDMStructureCTRegistrar {
 
 	@Activate
 	public void activate() {
-		_ctConfigurationRegistrar.register(
+		_ctDefinitionRegistrar.register(
 			_builder.setContentType(
-				"Dynamic Data Mapping Template"
+				"Dynamic Data Mapping Structure"
 			).setContentTypeLanguageKey(
-				"dynamic-data-mapping-template"
+				"dynamic-data-mapping-structure"
 			).setEntityClasses(
-				DDMTemplate.class, DDMTemplateVersion.class
+				DDMStructure.class, DDMStructureVersion.class
 			).setResourceEntitiesByCompanyIdFunction(
-				this::_fetchDDMTemplates
+				this::_fetchDDMStructures
 			).setResourceEntityByResourceEntityIdFunction(
-				_ddmTemplateLocalService::fetchTemplate
+				_ddmStructureLocalService::fetchStructure
 			).setEntityIdsFromResourceEntityFunctions(
-				DDMTemplate::getTemplateId, this::_fetchLatestTemplateVersionId
+				DDMStructure::getStructureId,
+				this::_fetchLatestStructureVersionId
 			).setVersionEntitiesFromResourceEntityFunction(
-				ddmTemplate ->
-					_ddmTemplateVersionLocalService.getTemplateVersions(
-						ddmTemplate.getTemplateId())
+				ddmStructure ->
+					_ddmStructureVersionLocalService.getStructureVersions(
+						ddmStructure.getStructureId())
 			).setVersionEntityByVersionEntityIdFunction(
-				_ddmTemplateVersionLocalService::fetchDDMTemplateVersion
+				_ddmStructureVersionLocalService::fetchDDMStructureVersion
 			).setVersionEntityDetails(
 				null, CTFunctions.getFetchSiteNameFunction(),
 				ddmStructureVersion -> ddmStructureVersion.getName(
 					LocaleUtil.getMostRelevantLocale()),
-				DDMTemplateVersion::getVersion
+				DDMStructureVersion::getVersion
 			).setEntityIdsFromVersionEntityFunctions(
-				DDMTemplateVersion::getTemplateId,
-				DDMTemplateVersionModel::getTemplateVersionId
+				DDMStructureVersion::getStructureId,
+				DDMStructureVersion::getStructureVersionId
 			).setVersionEntityStatusInfo(
-				new Integer[] {WorkflowConstants.STATUS_APPROVED},
-				ddmTemplateVersion -> WorkflowConstants.STATUS_APPROVED
+				new Integer[] {
+					WorkflowConstants.STATUS_APPROVED,
+					WorkflowConstants.STATUS_DRAFT
+				},
+				DDMStructureVersion::getStatus
 			).build());
 	}
 
-	private List<DDMTemplate> _fetchDDMTemplates(long companyId) {
-		DynamicQuery dynamicQuery = _ddmTemplateLocalService.dynamicQuery();
+	private List<DDMStructure> _fetchDDMStructures(long companyId) {
+		DynamicQuery dynamicQuery = _ddmStructureLocalService.dynamicQuery();
 
 		Property companyIdProperty = PropertyFactoryUtil.forName("companyId");
 
 		dynamicQuery.add(companyIdProperty.eq(companyId));
 
-		return _ddmTemplateLocalService.dynamicQuery(dynamicQuery);
+		return _ddmStructureLocalService.dynamicQuery(dynamicQuery);
 	}
 
-	private Serializable _fetchLatestTemplateVersionId(
-		DDMTemplate ddmTemplate) {
+	private Serializable _fetchLatestStructureVersionId(
+		DDMStructure ddmStructure) {
 
 		try {
-			DDMTemplateVersion ddmTemplateVersion =
-				ddmTemplate.getLatestTemplateVersion();
+			DDMStructureVersion ddmStructureVersion =
+				ddmStructure.getStructureVersion();
 
-			return ddmTemplateVersion.getTemplateVersionId();
+			return ddmStructureVersion.getStructureVersionId();
 		}
 		catch (PortalException pe) {
 			if (_log.isDebugEnabled()) {
-				_log.debug("Unable to get template version ID", pe);
+				_log.debug(
+					"An error occured when getting structure version ID", pe);
 			}
 
 			return 0;
@@ -111,21 +114,18 @@ public class DDMTemplateCTConfigurationRegistrar {
 	}
 
 	private static final Log _log = LogFactoryUtil.getLog(
-		DDMTemplateCTConfigurationRegistrar.class);
+		DDMStructureCTRegistrar.class);
 
 	@Reference(scope = ReferenceScope.PROTOTYPE_REQUIRED)
-	private CTConfigurationBuilder<DDMTemplate, DDMTemplateVersion> _builder;
+	private CTDefinitionBuilder<DDMStructure, DDMStructureVersion> _builder;
 
 	@Reference
-	private CTConfigurationRegistrar _ctConfigurationRegistrar;
+	private CTDefinitionRegistrar _ctDefinitionRegistrar;
 
 	@Reference
-	private DDMTemplateLocalService _ddmTemplateLocalService;
+	private DDMStructureLocalService _ddmStructureLocalService;
 
 	@Reference
-	private DDMTemplateVersionLocalService _ddmTemplateVersionLocalService;
-
-	@Reference
-	private GroupLocalService _groupLocalService;
+	private DDMStructureVersionLocalService _ddmStructureVersionLocalService;
 
 }
